api/requirements.txt
google-generativeai 
requests 
textblob
sentence_transformers
spacy
fastapi
uvicorn
pydantic
beautifulsoup4